Updated nightly builds code

+++ tools/nightly_build.sh
@@ -1,8 +1,11 @@
-#!/bin/bash 
+#!/bin/bash  -x
 
 ROOT_PATH=`dirname $0`
 BSD_PATH=`dirname $0`/bsd-build.sh
 SOLARIS_PATH=`dirname $0`/solaris-build.sh
+
+EXCLUDE_TEST=test_locale_icu_vs_os_timezone
+#EXCLUDE_TEST=
 
 cd /tmp
 rm -fr /tmp/faults
@@ -13,13 +16,16 @@ rm -f report.txt
 FLAGS=
 
 rm -fr /tmp/nb
-svn export svn://svn.code.sf.net/p/cppcms/code/framework/trunk nb
-cd nb
-tar -xjf cppcms_boost.tar.bz2
+
+source $ROOT_PATH/url.sh
+REPO_REV=`svn info $REPO_URL | grep Revision`
+svn export $REPO_URL nb || exit 1
+rm -f nb.tar.gz
+tar -czvf nb.tar.gz nb
 
 cd /tmp
 
-gcc_44_stlport()
+clang_38_libcpp()
 {
 	cd /tmp/nb
 	rm -fr build
@@ -27,7 +33,7 @@ gcc_44_stlport()
 	cd build
 
 
-	if cmake $FLAGS -DUSE_STLPORT=ON -DDISABLE_ICU_LOCALE=ON -DDISABLE_STD_LOCALE=ON -DCMAKE_BUILD_TYPE=Debug -DDISABLE_STATIC=ON .. && make && make test
+	if cmake -DCMAKE_C_COMPILER=/usr/bin/clang -DCMAKE_CXX_COMPILER=/usr/bin/clang++ -DCMAKE_CXX_FLAGS="-stdlib=libc++ -I /usr/include/libcxxabi" -DDISABLE_STATIC=ON -DDISABLE_ICU_LOCALE=ON .. && make -j 4 && ctest -E "$EXCLUDE_TEST"
 	then
 		return 0;
 	else
@@ -39,7 +45,7 @@ gcc_44_stlport()
 	fi
 
 }
-gcc_44()
+gcc_53()
 {
 	cd /tmp/nb
 	rm -fr build
@@ -47,7 +53,7 @@ gcc_44()
 	cd build
 
 
-	if cmake $FLAGS -DDISABLE_STATIC=ON .. && make && make test
+	if cmake $FLAGS -DDISABLE_STATIC=ON .. && make -j 4 && ctest -E "$EXCLUDE_TEST"
 	then
 		return 0;
 	else
@@ -71,7 +77,7 @@ gcc_45()
 	cd build
 
 
-	if cmake $FLAGS -DDISABLE_STATIC=ON -DCMAKE_BUILD_TYPE=Debug -DCMAKE_C_COMPILER=/opt/gcc45/bin/gcc-4.5 -DCMAKE_CXX_COMPILER=/opt/gcc45/bin/g++-4.5 .. && make && make test
+	if cmake $FLAGS -DDISABLE_STATIC=ON -DCMAKE_BUILD_TYPE=Debug -DCMAKE_C_COMPILER=/opt/gcc45/bin/gcc-4.5 -DCMAKE_CXX_COMPILER=/opt/gcc45/bin/g++-4.5 .. && make && ctest -E "$EXCLUDE_TEST"
 	then
 		return 0;
 	else
@@ -95,7 +101,7 @@ clangcc()
 	cd build
 
 	if cmake $FLAGS -DCMAKE_C_COMPILER=`which clang` -DCMAKE_CXX_COMPILER=`which clang++` -DDISABLE_STATIC=ON -DCMAKE_BUILD_TYPE=Debug .. \
-		 && make && make test
+		 && make && ctest -E "$EXCLUDE_TEST"
 	then
 		return 0;
 	else
@@ -119,7 +125,7 @@ intel()
 	cd build
 
 	if cmake $FLAGS -DCMAKE_C_COMPILER=`which icc` -DCMAKE_CXX_COMPILER=`which icpc` -DDISABLE_STATIC=ON .. \
-		 && make && make test
+		 && make && ctest -E "$EXCLUDE_TEST"
 	then
 		return 0;
 	else
@@ -142,7 +148,7 @@ gcc_450x()
 	cd build
 
 
-	if cmake $FLAGS -DDISABLE_STATIC=ON -DCMAKE_BUILD_TYPE=Debug -DCMAKE_CXX_FLAGS:STRING=-std=c++0x  -DCMAKE_C_COMPILER=/opt/gcc45/bin/gcc-4.5 -DCMAKE_CXX_COMPILER=/opt/gcc45/bin/g++-4.5 .. && make && make test
+	if cmake $FLAGS -DDISABLE_STATIC=ON -DCMAKE_BUILD_TYPE=Debug -DCMAKE_CXX_FLAGS:STRING=-std=c++0x  -DCMAKE_C_COMPILER=/opt/gcc45/bin/gcc-4.5 -DCMAKE_CXX_COMPILER=/opt/gcc45/bin/g++-4.5 .. && make && ctest -E "$EXCLUDE_TEST"
 	then
 		return 0;
 	else
@@ -161,52 +167,80 @@ cd /tmp
 # LINUX 
 #####################
 
-for TEST in gcc_44 intel gcc_44_stlport clangcc
-do
-	FILE=/tmp/$TEST.txt
-	if $TEST &> $FILE 
-	then
-		echo $TEST - pass >>/tmp/report.txt
-	else
-		echo $TEST - fail >>/tmp/report.txt
-	fi
-	cp $FILE /tmp/faults/
-done
+if true
+then
+
+	for TEST in gcc_53 clang_38_libcpp
+	do
+		FILE=/tmp/$TEST.txt
+		if $TEST &> $FILE 
+		then
+			echo $TEST - pass >>/tmp/report.txt
+		else
+			echo $TEST - fail >>/tmp/report.txt
+		fi
+		cp $FILE /tmp/faults/
+	done
+
+fi
 
 #####################
 # FreeBSD
 #####################
 
-VBoxHeadless -s FreeBSD &
-PID=$!
-sleep 60
+if true
+then
 
-scp -P 2223 $BSD_PATH artik@localhost:/tmp
-ssh -p 2223 artik@localhost /tmp/bsd-build.sh
-scp -P 2223 artik@localhost:/tmp/report.txt /tmp/tmp.txt
-cat /tmp/tmp.txt >>/tmp/report.txt
-scp -P 2223 'artik@localhost:/tmp/faults/*' /tmp/faults/
+	VBoxHeadless -s FreeBSD &
+	PID=$!
+	sleep 60
 
-VBoxManage controlvm FreeBSD acpipowerbutton
-wait $PID
+	for x in 1..10 
+	do
+		echo trying to copy files
+		if scp -P 2223 $BSD_PATH /tmp/nb.tar.gz artik@localhost:/tmp
+		then
+			echo "Done"
+			break
+		else
+			sleep 10
+		fi
+	done
+
+	scp -P 2223 $ROOT_PATH/url.sh artik@localhost:/tmp
+	ssh -p 2223 artik@localhost /tmp/bsd-build.sh
+	scp -P 2223 artik@localhost:/tmp/report.txt /tmp/tmp.txt
+	cat /tmp/tmp.txt >>/tmp/report.txt
+	scp -P 2223 'artik@localhost:/tmp/faults/*' /tmp/faults/
+
+	VBoxManage controlvm FreeBSD acpipowerbutton
+	wait $PID
+
+fi
 
 
 #####################
 # Solaris
 #####################
 
-VBoxHeadless -s Solaris &
-PID=$!
-sleep 300
+if true
+then
 
-scp -P 2222 $SOLARIS_PATH artik@localhost:/tmp
-ssh -p 2222 artik@localhost /tmp/solaris-build.sh
-scp -P 2222 artik@localhost:/tmp/report.txt /tmp/tmp.txt
-cat /tmp/tmp.txt >>/tmp/report.txt
-scp -P 2222 'artik@localhost:/tmp/faults/*' /tmp/faults/
+	VBoxHeadless -s Solaris &
+	PID=$!
+	sleep 300
 
-VBoxManage controlvm Solaris acpipowerbutton
-wait $PID
+	scp -P 2222 $SOLARIS_PATH /tmp/nb.tar.gz artik@localhost:/tmp
+	scp -P 2222 $ROOT_PATH/url.sh artik@localhost:/tmp
+	ssh -p 2222 artik@localhost /tmp/solaris-build.sh
+	scp -P 2222 artik@localhost:/tmp/report.txt /tmp/tmp.txt
+	cat /tmp/tmp.txt >>/tmp/report.txt
+	scp -P 2222 'artik@localhost:/tmp/faults/*' /tmp/faults/
+
+	VBoxManage controlvm Solaris acpipowerbutton
+	wait $PID
+
+fi
 
 #####################
 # Debian Armel
@@ -218,45 +252,48 @@ wait $PID
 # Windows
 #####################
 
-cd /tmp
-rm -fr nb/build
-rm -f nb.tar.gz
-tar -czf nb.tar.gz nb
-cp $ROOT_PATH/mingw-build.sh .
-smbclient -U guest -N '//192.168.2.100/storage' -c 'put nb.tar.gz'
-smbclient -U guest -N '//192.168.2.100/storage' -c 'put mingw-build.sh'
-smbclient -U guest -N '//192.168.2.100/storage' -c 'del faults.tar.gz'
-
-VBoxHeadless -s XP &
-PID=$!
-
-for x in {1..240}
-do
-	if [ -e /mnt/raid/storage/faults.tar.gz ] 
-	then
-		sleep 1
-		break
-	else
-		sleep 30
-	fi
-done
-
-VBoxManage controlvm XP acpipowerbutton
-wait $PID
-
-if [ -e /mnt/raid/storage/faults.tar.gz ] 
+if true
 then
-	mkdir temp
-	cp /mnt/raid/storage/faults.tar.gz temp
-	cd temp
-	tar -xzf faults.tar.gz
-	cat faults/report.txt >>/tmp/report.txt
-	rm faults/report.txt
-	cp faults/* /tmp/faults/
-	cd ..
-	rm -fr temp 
-fi
 
+	cd /tmp
+	rm -fr nb/build
+	rm -f nb.tar.gz
+	tar -czf nb.tar.gz nb
+	cp $ROOT_PATH/mingw-build.sh .
+	smbclient -U guest -N '//192.168.2.100/storage' -c 'put nb.tar.gz'
+	smbclient -U guest -N '//192.168.2.100/storage' -c 'put mingw-build.sh'
+	smbclient -U guest -N '//192.168.2.100/storage' -c 'del faults.tar.gz'
+
+	VBoxHeadless -s XP &
+	PID=$!
+
+	for x in {1..240}
+	do
+		if [ -e /mnt/raid1/storage/faults.tar.gz ] 
+		then
+			sleep 1
+			break
+		else
+			sleep 30
+		fi
+	done
+
+	VBoxManage controlvm XP acpipowerbutton
+	wait $PID
+
+	if [ -e /mnt/raid1/storage/faults.tar.gz ] 
+	then
+		mkdir temp
+		cp /mnt/raid1/storage/faults.tar.gz temp
+		cd temp
+		tar -xzf faults.tar.gz
+		cat faults/report.txt >>/tmp/report.txt
+		rm faults/report.txt
+		cp faults/* /tmp/faults/
+		cd ..
+		rm -fr temp 
+	fi
+fi
 
 #####################
 # Report
@@ -264,17 +301,14 @@ fi
 
 
 cp /tmp/report.txt /tmp/faults
-cp /tmp/report.txt /home/artik/vserver-www/www/files/nightly-build-report.txt
+cp /tmp/report.txt /home/artik/nightly-build-report.txt
 
 cd /tmp
 rm -fr /tmp/nightly-build-report
 mv /tmp/faults /tmp/nightly-build-report
 
 tar -czvf nightly-build-report.tar.gz nightly-build-report
-cp nightly-build-report.tar.gz ~/
-cp nightly-build-report.tar.gz /home/artik/vserver-www/www/files/
-cp -a nightly-build-report /home/artik/vserver-www/www/files/
-$ROOT_PATH/report_to_html.py < ~/vserver-www/www/files/nightly-build-report.txt >~/vserver-www/www/files/nightly-build-report.html
+$ROOT_PATH/report_to_html.py "$REPO_URL" "$REPO_REV" < /home/artik/nightly-build-report.txt >/home/artik/nightly-build-report.html
 
 $HOME/bin/sync_build.sh
